refactor(backend): remove app "alerts" and add new app "runs"
