--- backend/alerts/multicast.py
@@ -1,2 +1,0 @@
-def test_signal(sender, **kwargs):
-    print("Vamos a ver cómo está la cosa...")
